Breakout cryptographic functions
